Commented out the running of MCell and print command instead.
The instructions run MCell as a separate step after running mdlr2mdl.py.
But the current version of mdlr2mdl.py actually run MCell as a final
step. To make the code match the instructions (and to better separate
file creation from running), the call(cmd) line was commented out and
replaced with a printing of the command (for easy copy and paste).