Moved viz output to end of main.mdl, but it ends up before mols etc.
The MDLR to MDL process appears to force certain sections at the end of
the generated MDL. These include molecules, reactions, seed(species),
surface classes and regions. So placing the viz block at the end of the
MDLR file doesn't put it at the end of the generated MDL file.